[#2869] Move all the result framework related models to a submodule
Use`result` submodule, instead of `indicator` submodule.
