+++ akvo/iati/imports/mappers/results.py
@@ -4,12 +4,13 @@
 # See more details in the license.txt file located at the root folder of the Akvo RSR module.
 # For additional details on the GNU license please see < http://www.gnu.org/licenses/agpl.html >.
 
-from ....rsr.models.indicator import (Indicator, IndicatorPeriod, IndicatorPeriodActualDimension,
-                                      IndicatorPeriodActualLocation, IndicatorPeriodTargetDimension,
-                                      IndicatorPeriodTargetLocation, IndicatorReference)
-from ....rsr.models.result import Result
-
 from .. import ImportMapper
+from akvo.rsr.models.result import (Indicator, IndicatorPeriod,
+                                    IndicatorPeriodActualDimension,
+                                    IndicatorPeriodActualLocation,
+                                    IndicatorPeriodTargetDimension,
+                                    IndicatorPeriodTargetLocation,
+                                    IndicatorReference, Result)
 
 
 class Results(ImportMapper):
