+++ akvo/rsr/models/__init__.py
@@ -39,13 +39,13 @@ from .iati_export import IatiExport
 from .iati_import import IatiImport
 from .iati_import_job import IatiImportJob, CordaidZipIatiImportJob
 from .iati_import_log import IatiImportLog
-from .indicator import (Disaggregation, Indicator, IndicatorDimension,
-                        IndicatorLabel, IndicatorPeriod, IndicatorPeriodData,
-                        IndicatorPeriodDataComment, IndicatorReference,
-                        IndicatorPeriodActualDimension,
-                        IndicatorPeriodTargetDimension,
-                        IndicatorPeriodActualLocation,
-                        IndicatorPeriodTargetLocation)
+from .result import (Disaggregation, Indicator, IndicatorDimension,
+                     IndicatorLabel, IndicatorPeriod, IndicatorPeriodData,
+                     IndicatorPeriodDataComment, IndicatorReference,
+                     IndicatorPeriodActualDimension,
+                     IndicatorPeriodTargetDimension,
+                     IndicatorPeriodActualLocation,
+                     IndicatorPeriodTargetLocation)
 from .internal_organisation_id import InternalOrganisationID
 from .keyword import Keyword
 from .legacy_data import LegacyData
