+++ akvo/rsr/models/result/__init__.py
@@ -1,0 +1,35 @@
+# -*- coding: utf-8 -*-
+
+# Akvo RSR is covered by the GNU Affero General Public License.
+# See more details in the license.txt file located at the root folder of the Akvo RSR module.
+# For additional details on the GNU license please see < http://www.gnu.org/licenses/agpl.html >.
+
+from .disaggregation import Disaggregation
+from .indicator import Indicator
+from .indicator_dimension import IndicatorDimension
+from .indicator_label import IndicatorLabel
+from .indicator_period import IndicatorPeriod
+from .indicator_period_actual_dimension import IndicatorPeriodActualDimension
+from .indicator_period_actual_location import IndicatorPeriodActualLocation
+from .indicator_period_data import IndicatorPeriodData
+from .indicator_period_data_comment import IndicatorPeriodDataComment
+from .indicator_period_target_dimension import IndicatorPeriodTargetDimension
+from .indicator_period_target_location import IndicatorPeriodTargetLocation
+from .indicator_reference import IndicatorReference
+from .result import Result
+
+__all__ = [
+    'Disaggregation',
+    'Indicator',
+    'IndicatorDimension',
+    'IndicatorLabel',
+    'IndicatorPeriod',
+    'IndicatorPeriodActualDimension',
+    'IndicatorPeriodActualLocation',
+    'IndicatorPeriodData',
+    'IndicatorPeriodDataComment',
+    'IndicatorPeriodTargetDimension',
+    'IndicatorPeriodTargetLocation',
+    'IndicatorReference',
+    'Result',
+]
