+++ akvo/rsr/models/result/indicator_label.py
