+++ akvo/rsr/models/result/indicator_period_actual_location.py
