+++ akvo/rsr/models/result/indicator_period_target_location.py
